requirements.txt
Django==4.2.6
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
